+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CORS error due to incorrect configuration when executing the SQL query
 - Corrected redirection to a page from a link in an email when the token has expired 
 - Activity tab: search functionality when switching process statuses
+### Changed
+- The dataset schema has been converted to ODF format 
 - Flows tab: the status filter was incorrectly initialized for request